Image Optimization
Image compression from squoosh.app in 2 formats (WebP (modern format) and OxiPng), and resizing. Additionally, the use of the picture tag provides two different sources, one for each format, in order to let each browser choose which format can render. Although in the img tag the source entails the .png, the actual image will be of the .webp extension whenever it is supported, which is both lighter and of better quality.